django-part/requirements.txt
asgiref==3.3.1
Django==3.1.3
django-cors-headers==3.5.0
django-nested-admin==3.3.2
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
Pillow==8.1.2
PyJWT==1.7.1
python-monkey-business==1.0.0
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
